Lead Navigator AI Frontend/requirements.txt
dash==2.18.2
dash-bootstrap-components==1.6.0
plotly==5.24.1
pandas==2.2.3
polars==1.9.0
openpyxl==3.1.5
chardet==5.2.0
python-dateutil==2.9.0.post0
requests==2.32.3
python-dotenv==1.0.1
PyJWT==2.9.0
gunicorn==23.0.0